refactor: Port 404 page to next13
